requirements.txt
requests
pandas
sqlalchemy
pymysql
openai>=1.97.1
matplotlib
seaborn
streamlit
wordcloud
scikit-learn
python-dotenv
